requirements.txt
python-telegram-bot>=13.0
Flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.15.0